\name{energymethod}
\alias{energymethod}
\docType{package}
\title{Energy Method}
\description{
Given two samples from a multi-channel functional distribtution, this package implements the energy method to perform a test of equality of mean. It returns channel-wise p-values and the global p-value. 
}

\author{
  David Colin Decker \email{d.colin.decker@gmail.com}
}

